Added the ability for the extension to give error messages and status when client connects.
Although very rare, the extension might not accept the new client and
instead close the connection. The client should be able to know why it
cannot connect to know what it should do next.

- Added ConnectionStatus event.
- Extension versions (api and general version) are removed from BridgeInfo and instead included in the ConnectionStatus event since it is static and does not changed at runtime.
- BridgeInfo event is included in the ConnectionStatus event, but can also be sent on its own if updated at runtime.

The new event is considered final and should never changed in any
version to keep compatibility.